Simplify crawling view
