+++ src/ui/scrapbook.rs
@@ -1,26 +1,24 @@
 use std::fmt::Write;
 
-use ahash::HashSet;
 use chrono::Local;
 use iced::{
     alignment::Horizontal,
     theme,
     widget::{
-        button, checkbox, column, horizontal_space, pick_list, progress_bar,
-        row, scrollable, text, vertical_space, Image,
+        button, checkbox, column, horizontal_space, row, scrollable, text,
+        vertical_space, Image,
     },
     Alignment, Element, Length,
 };
 use iced_aw::number_input;
 use num_format::ToFormattedString;
 
-use super::remaining_minutes;
+use super::{remaining_minutes, view_crawling};
 use crate::{
     config::Config,
-    crawler::CrawlingOrder,
     message::Message,
     player::{AccountInfo, AccountStatus},
-    server::{CrawlingStatus, ServerInfo},
+    server::ServerInfo,
     ClassImages,
 };
 
@@ -154,117 +152,8 @@ pub fn view_scrapbook<'a>(
         left_col = left_col.push(scrollable(log).height(Length::Fixed(200.0)));
     }
     left_col = left_col.push(vertical_space());
-    let sid = server.ident.id;
-    let mut banned = HashSet::default();
 
-    match &server.crawling {
-        CrawlingStatus::Crawling {
-            threads,
-            que,
-            player_info,
-            ..
-        } => {
-            let lock = que.lock().unwrap();
-            let remaining = lock.count_remaining();
-            let crawled = player_info.len();
-            let total = remaining + crawled;
-
-            banned = lock.invalid_accounts.iter().cloned().collect();
-
-            let progress_text = text(format!(
-                "Fetched {}/{}",
-                crawled.to_formatted_string(&config.num_format),
-                total.to_formatted_string(&config.num_format)
-            ));
-            left_col = left_col.push(progress_text);
-
-            let progress = progress_bar(0.0..=total as f32, crawled as f32)
-                .height(Length::Fixed(10.0));
-            left_col = left_col.push(progress);
-
-            let thread_num =
-                number_input(*threads, config.max_threads, move |nv| {
-                    Message::CrawlerSetThreads {
-                        server: sid,
-                        new_count: nv,
-                    }
-                });
-            let thread_num =
-                row!(text("Threads: "), horizontal_space(), thread_num)
-                    .align_items(Alignment::Center);
-            left_col = left_col.push(thread_num);
-            let order_picker = pick_list(
-                [
-                    CrawlingOrder::Random,
-                    CrawlingOrder::TopDown,
-                    CrawlingOrder::BottomUp,
-                ],
-                Some(lock.order),
-                |nv| Message::OrderChange {
-                    server: server.ident.id,
-                    new: nv,
-                },
-            );
-            left_col = left_col.push(
-                row!(
-                    text("Crawling Order:").width(Length::FillPortion(1)),
-                    order_picker.width(Length::FillPortion(1))
-                )
-                .align_items(Alignment::Center),
-            );
-
-            if config.show_crawling_restrict
-                || !lock.lvl_skipped_accounts.is_empty()
-            {
-                let old_max = lock.max_level;
-                let old_min = lock.min_level;
-
-                let set_min_lvl =
-                    number_input(lock.min_level, 9999u32, move |nv| {
-                        Message::CrawlerSetMinMax {
-                            server: sid,
-                            min: nv,
-                            max: old_max,
-                        }
-                    });
-                let thread_num =
-                    row!(text("Min Lvl: "), horizontal_space(), set_min_lvl)
-                        .align_items(Alignment::Center);
-                left_col = left_col.push(thread_num);
-
-                let set_min_lvl =
-                    number_input(lock.max_level, 9999u32, move |nv| {
-                        Message::CrawlerSetMinMax {
-                            server: sid,
-                            min: old_min,
-                            max: nv,
-                        }
-                    });
-                let thread_num =
-                    row!(text("Max Lvl: "), horizontal_space(), set_min_lvl)
-                        .align_items(Alignment::Center);
-                left_col = left_col.push(thread_num);
-            }
-
-            let clear = button("Clear HoF").on_press(Message::ClearHof(sid));
-            let save = button("Save HoF").on_press(Message::SaveHoF(sid));
-            left_col = left_col.push(
-                column!(row!(clear, save).spacing(10))
-                    .align_items(Alignment::Center),
-            );
-
-            drop(lock);
-        }
-        CrawlingStatus::Waiting => {
-            left_col = left_col.push(text("Waiting for Player..."));
-        }
-        CrawlingStatus::Restoring => {
-            left_col = left_col.push(text("Loading Server Data..."));
-        }
-        CrawlingStatus::CrawlingFailed(_) => {
-            left_col = left_col.push(text("Crawling Failed"));
-        }
-    }
+    left_col = left_col.push(view_crawling(server, config));
 
     let mut name_bar = column!();
     name_bar = name_bar.push(row!(
@@ -288,9 +177,6 @@ pub fn view_scrapbook<'a>(
 
     let mut target_list = column!().spacing(10);
     for v in &si.best {
-        if banned.contains(&v.info.name) {
-            continue;
-        }
         let mut target_ident = row!()
             .align_items(Alignment::Start)
             .spacing(5)
